feat: renaming entities to models for clarity
