+++ models/classroom.go
@@ -1,4 +1,4 @@
-package entities
+package models
 
 import "github.com/google/uuid"
